#  Succeed: use private properties lookup instead of calling getters - gains ~6%